[minor] Move create_executor() function to separate module
